refactor(workflow): unify WorkflowPanel with shared WorkflowForm
Extract shared execution/history components and draft conversion so AI panel
and workflow window reuse the same form and result rendering (~1000 lines removed).
